llm_test/requirements_gpu.txt
torch==2.7.1+cu118
transformers==5.1.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
datasets>=2.0.0
pyyaml>=6.0
pandas>=1.5.0
numpy>=1.21.0